refactor(vscode): replace client-side routing with server-side filtering
Both language servers now receive every open Prisma document through
standard LSP document synchronization and decide locally, from the
`// use prisma-next` directive, whether to respond. The legacy server
already filtered marked documents; the Prisma Next server does the
mirror image since prisma@8.0.0-rc.8-dev.2.

This makes the client-side ownership-transfer machinery unnecessary:
the ownership coordinator, routing transitions, gating middlewares,
and lifecycle controller are deleted. The extension now only decides
which servers to start: at most one legacy server per window, and at
most one Prisma Next server per workspace folder, started the first
time an open document under that folder carries the directive.
