+++ packages/vscode/src/plugins/prisma-language-server/index.ts
@@ -2,8 +2,27 @@ import path from 'path'
 import os from 'os'
 import minimatch from 'minimatch'
 
-import { commands, ExtensionContext, TextDocument, window, workspace, languages, WorkspaceConfiguration } from 'vscode'
-import { LanguageClientOptions } from 'vscode-languageclient'
+import {
+  CancellationToken,
+  CodeAction,
+  CodeActionContext,
+  Command,
+  commands,
+  ExtensionContext,
+  Range,
+  TextDocument,
+  window,
+  workspace,
+  languages,
+  WorkspaceConfiguration,
+} from 'vscode'
+import {
+  CodeAction as lsCodeAction,
+  CodeActionParams,
+  CodeActionRequest,
+  LanguageClientOptions,
+  ProvideCodeActionsSignature,
+} from 'vscode-languageclient'
 import { LanguageClient, ServerOptions, TransportKind } from 'vscode-languageclient/node'
 import TelemetryReporter from '../../telemetryReporter'
 import {
@@ -11,6 +30,7 @@ import {
   checkForMinimalColorTheme,
   checkForOtherPrismaExtension,
   isDebugOrTestSession,
+  isPrismaNextSchema,
   isSnippetEdit,
   restartClient,
   createLegacyLanguageServer,
@@ -21,40 +41,26 @@ import FileWatcher from 'watcher'
 import { CodelensProvider, generateClient } from '../../CodeLensProvider'
 import * as prisma6Handling from '../../prisma6Handling'
 import { getPackageJSON } from '../../getPackageJSON'
-import { DocumentOwnershipCoordinator } from './documentOwnership'
-import { createLegacyClientMiddleware, type LegacyClientMiddleware } from './legacyClientMiddleware'
-import { createPrepareDocumentRoutingCommit } from './documentRouting'
-import { PrismaNextClientRegistry } from './prismaNextClientRegistry'
-import { LanguageServerLifecycleController } from './languageServerLifecycle'
+import { PrismaNextClients } from './prismaNextClients'
 
-let legacyClient: LanguageClient
-let legacyServerModule: string
+let client: LanguageClient
+let serverModule: string
 let telemetry: TelemetryReporter
 let fileWatcher: FileWatcher.type | undefined
-let languageServerLifecycle: LanguageServerLifecycleController | undefined
-let prismaNextClientRegistry: PrismaNextClientRegistry | undefined
+let prismaNextClients: PrismaNextClients | undefined
 
 const isDebugMode = () => process.env.VSCODE_DEBUG_MODE === 'true'
-const logLegacyClientError = (error: unknown): void => {
-  console.error('Legacy Prisma Language Server failed', error)
-}
 
-const activateLegacyClientNow = async (
-  context: ExtensionContext,
-  legacyClientOptions: LanguageClientOptions,
-  lifecycle: LanguageServerLifecycleController,
-): Promise<void> => {
-  lifecycle.assertActive()
+const activateClient = (context: ExtensionContext, clientOptions: LanguageClientOptions) => {
   const prismaConfig = workspace.getConfiguration('prisma')
-  lifecycle.assertActive()
-  const client = createLegacyLanguageServer(getLegacyServerOptions(prismaConfig, context), legacyClientOptions)
-  lifecycle.publishLegacyClient(client)
-  legacyClient = client
+  // Create the language client
+  const serverOptions = getServerOptions(prismaConfig, context)
+  client = createLegacyLanguageServer(serverOptions, clientOptions)
 
-  lifecycle.assertActive()
-  context.subscriptions.push(client.start())
-  await lifecycle.waitFor(client.onReady())
-  lifecycle.assertActive()
+  const disposable = client.start()
+
+  // Start the client. This will also launch the server
+  context.subscriptions.push(disposable)
 }
 
 const onFileChange = (filepath: string) => {
@@ -119,164 +125,109 @@ const plugin: PrismaVSCodePlugin = {
 
     setGenerateWatcher(!!workspace.getConfiguration('prisma').get('fileWatcher'))
 
-    let legacyUnavailable = false
-    let restarting = false
-    languageServerLifecycle?.dispose().catch(logLegacyClientError)
-    const lifecycle = new LanguageServerLifecycleController()
-    languageServerLifecycle = lifecycle
-    const ownership: DocumentOwnershipCoordinator = new DocumentOwnershipCoordinator({
-      workspace,
-      policy: {
-        isPinnedToPrisma6: () => !!workspace.getConfiguration('prisma').get<boolean>('pinToPrisma6'),
-        isLegacyUnavailable: () => legacyUnavailable,
-      },
-      prepareTransition: createPrepareDocumentRoutingCommit({
-        getOwnership: (): DocumentOwnershipCoordinator => ownership,
-        isActive: () => lifecycle.isActive,
-        isDocumentOpen: (document) => workspace.textDocuments.includes(document),
-        getLegacy: () => legacyClientMiddleware,
-        getPrismaNext: () => prismaNextClients,
-      }),
-    })
-    const prismaNextClients = new PrismaNextClientRegistry({
-      workspace,
-      ownership,
-      isActive: () => lifecycle.isActive,
-      getDocument: (uri) => workspace.textDocuments.find((document) => document.uri.toString() === uri.toString()),
-      createClient: (id, name, serverOptions, prismaNextClientOptions) =>
-        new LanguageClient(id, name, serverOptions, prismaNextClientOptions),
-      registerDisposable: (disposable) => context.subscriptions.push(disposable),
-      handleStartError: (workspaceFolder, error) => {
-        console.error(`Failed to start Prisma Next Language Server for ${workspaceFolder.uri.toString()}`, error)
-      },
-    })
-    prismaNextClientRegistry = prismaNextClients
-
-    const legacyClientMiddleware: LegacyClientMiddleware = createLegacyClientMiddleware({
-      ownership,
-      isActive: () => lifecycle.isActive,
-      getClient: () => legacyClient,
-      getDocument: (uri) => workspace.textDocuments.find((document) => document.uri.toString() === uri.toString()),
-      handleDiagnosticMessage: (message) => {
-        void prisma6Handling.handleDiagnostic(message, context)
-      },
-      isSnippetEdit,
-    })
+    // Both language servers receive every open Prisma document and each decides from the
+    // `// use prisma-next` directive in the document content whether to respond: the legacy
+    // server ignores documents with the directive, the Prisma Next server (prisma >=
+    // 8.0.0-rc.8-dev.2) ignores documents without it. The extension only decides which
+    // servers to start.
+    const nextClients = new PrismaNextClients((disposable) => context.subscriptions.push(disposable))
+    prismaNextClients = nextClients
 
     // Options to control the language client
-    const legacyClientOptions: LanguageClientOptions = {
+    const clientOptions: LanguageClientOptions = {
+      // Register the server for prisma documents
       documentSelector: [{ scheme: 'file', language: 'prisma' }],
-      middleware: legacyClientMiddleware,
-    }
 
-    let startRequested = false
-    let started = false
-    let ready = false
-    const needsLegacyLanguageServer = (document: TextDocument): boolean =>
-      document.languageId === 'prisma' && ownership.getDesiredOwner(document).kind === 'legacy'
-    const getOpenPrismaDocuments = (): TextDocument[] =>
-      workspace.textDocuments.filter((document) => document.languageId === 'prisma')
-
-    const waitForOwnershipOperations = async (operations: readonly Promise<unknown>[]): Promise<void> => {
-      const results = await Promise.allSettled(operations)
-      const failure = results.find((result): result is PromiseRejectedResult => result.status === 'rejected')
-      if (failure) throw failure.reason
-    }
-
-    const synchronizeDocuments = (documents: readonly TextDocument[]): Promise<void> => {
-      lifecycle.assertActive()
-      return waitForOwnershipOperations(documents.map((document) => ownership.synchronize(document)))
-    }
-
-    const invalidateLegacyDocuments = (): Promise<void> => {
-      lifecycle.assertActive()
-      const legacyDocuments = getOpenPrismaDocuments().filter(
-        (document) => ownership.getSettledOwner(document.uri).kind === 'legacy',
-      )
-      return waitForOwnershipOperations(legacyDocuments.map((document) => ownership.close(document)))
-    }
-
-    const startLegacyLanguageServerNow = async (): Promise<void> => {
-      lifecycle.assertActive()
-      if (started) return
-
-      started = true
-      await activateLegacyClientNow(context, legacyClientOptions, lifecycle)
-      ready = true
-      lifecycle.assertActive()
-      await synchronizeDocuments(getOpenPrismaDocuments())
-    }
-
-    const maybeStart = (document?: TextDocument): void => {
-      if (!lifecycle.isActive || startRequested || started) return
-      if (document ? !needsLegacyLanguageServer(document) : !workspace.textDocuments.some(needsLegacyLanguageServer))
-        return
-
-      startRequested = true
-      void lifecycle.enqueue(startLegacyLanguageServerNow).catch(logLegacyClientError)
-    }
-
-    const synchronizeDocument = (document: TextDocument): void => {
-      if (!lifecycle.isActive || restarting || document.languageId !== 'prisma') return
-      if (ownership.getDesiredOwner(document).kind === 'legacy') {
-        if (!started) maybeStart(document)
-        if (!ready) return
-      }
-      void ownership.synchronize(document).catch(logLegacyClientError)
-    }
-
-    const restartLanguageServerNow = async (): Promise<void> => {
-      lifecycle.assertActive()
-      if (!started) {
-        if (getOpenPrismaDocuments().some(needsLegacyLanguageServer)) {
-          await startLegacyLanguageServerNow()
-        } else {
-          await synchronizeDocuments(getOpenPrismaDocuments())
-        }
-        return
-      }
-
-      restarting = true
-      ready = false
-      legacyUnavailable = true
-      try {
-        try {
-          await invalidateLegacyDocuments()
-        } catch (error) {
-          legacyUnavailable = false
-          if (lifecycle.isActive) {
-            try {
-              await synchronizeDocuments(getOpenPrismaDocuments())
-            } catch (restoreError) {
-              logLegacyClientError(restoreError)
-            }
+      /* This middleware is part of the workaround for https://github.com/prisma/language-tools/issues/311 */
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+      middleware: {
+        async provideCodeActions(
+          document: TextDocument,
+          range: Range,
+          context: CodeActionContext,
+          token: CancellationToken,
+          _: ProvideCodeActionsSignature,
+        ) {
+          const params: CodeActionParams = {
+            textDocument: client.code2ProtocolConverter.asTextDocumentIdentifier(document),
+            range: client.code2ProtocolConverter.asRange(range),
+            context: client.code2ProtocolConverter.asCodeActionContext(context),
           }
-          throw error
-        }
 
-        lifecycle.assertActive()
-        const serverOptions = getLegacyServerOptions(workspace.getConfiguration('prisma'), context)
-        const replacement = await restartClient(context, legacyClient, serverOptions, legacyClientOptions, {
-          assertActive: () => lifecycle.assertActive(),
-          waitFor: (operation) => lifecycle.waitFor(operation),
-          onClientStopped: () => legacyClientMiddleware.resetClientState(),
-          onClientCreated: (replacementClient) => {
-            lifecycle.publishLegacyClient(replacementClient)
-            legacyClient = replacementClient
-          },
-        })
-        lifecycle.assertActive()
-        legacyClient = replacement
-        ready = true
-        legacyUnavailable = false
-        await synchronizeDocuments(getOpenPrismaDocuments())
-      } finally {
-        legacyUnavailable = false
-        restarting = false
+          return client.sendRequest(CodeActionRequest.type, params, token).then(
+            (values) => {
+              if (values === null) return undefined
+              const result: (CodeAction | Command)[] = []
+              for (const item of values) {
+                if (lsCodeAction.is(item)) {
+                  const action = client.protocol2CodeConverter.asCodeAction(item)
+                  if (
+                    isSnippetEdit(item, client.code2ProtocolConverter.asTextDocumentIdentifier(document)) &&
+                    item.edit !== undefined
+                  ) {
+                    action.command = {
+                      command: 'prisma.applySnippetWorkspaceEdit',
+                      title: '',
+                      arguments: [action.edit],
+                    }
+                    action.edit = undefined
+                  }
+                  result.push(action)
+                } else {
+                  const command = client.protocol2CodeConverter.asCommand(item)
+                  result.push(command)
+                }
+              }
+              return result
+            },
+            (_) => undefined,
+          )
+        },
+        handleDiagnostics: (uri, diagnostics, next) => {
+          for (const diagnostic of diagnostics) {
+            void prisma6Handling.handleDiagnostic(diagnostic.message, context)
+          }
+          next(uri, diagnostics)
+        },
+      },
+    }
+
+    const isPinnedToPrisma6 = () => !!workspace.getConfiguration('prisma').get<boolean>('pinToPrisma6')
+
+    let started = false
+    const needsLegacyLanguageServer = (doc: TextDocument): boolean =>
+      doc.languageId === 'prisma' && (isPinnedToPrisma6() || !isPrismaNextSchema(doc.getText()))
+    const needsPrismaNextLanguageServer = (doc: TextDocument): boolean =>
+      doc.languageId === 'prisma' && !isPinnedToPrisma6() && isPrismaNextSchema(doc.getText())
+
+    const maybeStart = () => {
+      if (started) return
+      if (!workspace.textDocuments.some(needsLegacyLanguageServer)) return
+      started = true
+      activateClient(context, clientOptions)
+    }
+
+    const maybeStartPrismaNext = (document: TextDocument) => {
+      if (needsPrismaNextLanguageServer(document)) {
+        nextClients.ensureClientFor(document)
       }
     }
 
-    const restartLanguageServer = (): Promise<void> => lifecycle.enqueue(restartLanguageServerNow)
+    const synchronizeServers = (document?: TextDocument) => {
+      maybeStart()
+      for (const doc of document ? [document] : workspace.textDocuments) {
+        maybeStartPrismaNext(doc)
+      }
+    }
+
+    const restartLanguageServer = async () => {
+      await nextClients.stopAll()
+      if (started) {
+        const serverOptions = getServerOptions(workspace.getConfiguration('prisma'), context)
+        client = await restartClient(context, client, serverOptions, clientOptions)
+      }
+      synchronizeServers()
+    }
 
     context.subscriptions.push(
       // when the file watcher settings change, we need to ensure they are applied
@@ -316,47 +267,23 @@ const plugin: PrismaVSCodePlugin = {
         await prismaConfig.update('fileWatcher', false /* value */, false /* workspace */)
       }),
 
-      commands.registerCommand('prisma.pinWorkspaceToPrisma6', () =>
-        lifecycle.enqueue(async () => {
-          lifecycle.assertActive()
-          await workspace.getConfiguration('prisma').update('pinToPrisma6', true, false)
-          lifecycle.assertActive()
-          await restartLanguageServerNow()
-          lifecycle.assertActive()
-          void window.showInformationMessage('Pinned workspace to Prisma 6.')
-        }),
-      ),
+      commands.registerCommand('prisma.pinWorkspaceToPrisma6', async () => {
+        await workspace.getConfiguration('prisma').update('pinToPrisma6', true, false)
+        await restartLanguageServer()
+        void window.showInformationMessage('Pinned workspace to Prisma 6.')
+      }),
 
-      commands.registerCommand('prisma.unpinWorkspaceFromPrisma6', () =>
-        lifecycle.enqueue(async () => {
-          lifecycle.assertActive()
-          await workspace.getConfiguration('prisma').update('pinToPrisma6', false, false)
-          lifecycle.assertActive()
-          await restartLanguageServerNow()
-          lifecycle.assertActive()
-          void window.showInformationMessage('Unpinned workspace from Prisma 6.')
-        }),
-      ),
+      commands.registerCommand('prisma.unpinWorkspaceFromPrisma6', async () => {
+        await workspace.getConfiguration('prisma').update('pinToPrisma6', false, false)
+        await restartLanguageServer()
+        void window.showInformationMessage('Unpinned workspace from Prisma 6.')
+      }),
 
-      workspace.onDidOpenTextDocument((document) => {
-        maybeStart(document)
-        synchronizeDocument(document)
-      }),
-      workspace.onDidChangeTextDocument((event) => {
-        maybeStart(event.document)
-        synchronizeDocument(event.document)
-      }),
-      workspace.onDidCloseTextDocument((document) => {
-        if (lifecycle.isActive && document.languageId === 'prisma') {
-          void ownership.close(document).catch(logLegacyClientError)
-        }
-      }),
+      workspace.onDidOpenTextDocument((document) => synchronizeServers(document)),
+      workspace.onDidChangeTextDocument((event) => synchronizeServers(event.document)),
     )
 
-    maybeStart()
-    for (const document of workspace.textDocuments) {
-      synchronizeDocument(document)
-    }
+    synchronizeServers()
 
     if (!isDebugOrTest) {
       const packageJSON = getPackageJSON(context)
@@ -377,40 +304,37 @@ const plugin: PrismaVSCodePlugin = {
     checkForMinimalColorTheme()
   },
   deactivate: async () => {
-    const lifecycle = languageServerLifecycle
-    const prismaNextClients = prismaNextClientRegistry
-    languageServerLifecycle = undefined
-    prismaNextClientRegistry = undefined
+    const nextClients = prismaNextClients
+    prismaNextClients = undefined
 
-    const deactivations = [lifecycle?.dispose(), prismaNextClients?.dispose()].filter(
-      (deactivation): deactivation is Promise<void> => deactivation !== undefined,
-    )
-    const results = await Promise.allSettled(deactivations)
+    const results = await Promise.allSettled([nextClients?.dispose(), client?.stop()])
     for (const result of results) {
-      if (result.status === 'rejected') logLegacyClientError(result.reason)
+      if (result.status === 'rejected') {
+        console.error('Prisma language server deactivation failed', result.reason)
+      }
     }
 
-    if (legacyClient && !isDebugOrTestSession()) {
+    if (client && !isDebugOrTestSession()) {
       telemetry.dispose() // eslint-disable-line @typescript-eslint/no-floating-promises
     }
   },
 }
 
-function getLegacyServerOptions(prismaConfig: WorkspaceConfiguration, context: ExtensionContext): ServerOptions {
+function getServerOptions(prismaConfig: WorkspaceConfiguration, context: ExtensionContext): ServerOptions {
   const pinToPrisma6 = prismaConfig.get<boolean>('pinToPrisma6')
 
   if (pinToPrisma6) {
     console.log('Using legacy Prisma 6 Language Server')
-    legacyServerModule = context.asAbsolutePath(path.join('dist/prisma6-language-server/bin.js'))
+    serverModule = context.asAbsolutePath(path.join('dist/prisma6-language-server/bin.js'))
   } else if (isDebugMode()) {
     // Use the legacy Language Server from the source tree for debugging.
     console.log('Using legacy Language Server from filesystem')
-    legacyServerModule = context.asAbsolutePath(path.join('../../packages/language-server/dist/bin'))
+    serverModule = context.asAbsolutePath(path.join('../../packages/language-server/dist/bin'))
   } else {
     console.log('Using legacy Language Server')
-    legacyServerModule = context.asAbsolutePath(path.join('dist/language-server/bin.js'))
+    serverModule = context.asAbsolutePath(path.join('dist/language-server/bin.js'))
   }
-  console.log(`legacyServerModule: ${legacyServerModule}`)
+  console.log(`serverModule: ${serverModule}`)
 
   // The debug options for the server
   // --inspect=6009: runs the server in Node's Inspector mode so VS Code can attach to the server for debugging
@@ -422,9 +346,9 @@ function getLegacyServerOptions(prismaConfig: WorkspaceConfiguration, context: E
   // If the extension is launched in debug mode then the debug server options are used
   // Otherwise the run options are used
   return {
-    run: { module: legacyServerModule, transport: TransportKind.ipc },
+    run: { module: serverModule, transport: TransportKind.ipc },
     debug: {
-      module: legacyServerModule,
+      module: serverModule,
       transport: TransportKind.ipc,
       options: debugOptions,
     },
